suffix only if req, validate arch exists, quiet

+++ volume-backup.sh
@@ -6,12 +6,17 @@ usage() {
 }
 
 backup() {
-    tar -cvjf /backup/$ARCHIVE.tar.bz2 -C /volume ./
+    tar -cjf /backup/$ARCHIVE -C /volume ./
 }
 
 restore() {
+    if ! [ -e /backup/$ARCHIVE ]; then
+        echo "Archive file $ARCHIVE does not exist"
+        exit 1
+    fi
+
     rm -rf /volume/* /volume/..?* /volume/.[!.]*
-    tar -C /volume/ -xvjf /backup/$ARCHIVE.tar.bz2
+    tar -C /volume/ -xjf /backup/$ARCHIVE
 }
 
 if [ $# -ne 2 ]; then
@@ -19,7 +24,8 @@ if [ $# -ne 2 ]; then
 fi
 
 OPERATION=$1
-ARCHIVE=$2
+
+ARCHIVE=${2%%.tar.bz2}.tar.bz2
 
 case "$OPERATION" in
 "backup" )
